FLX-679 If you create a funding source, link a request transaction funding to it, then delete the funding source, you get an error

+++ lib/extensions/models/fluxx_request_transaction.rb
@@ -331,7 +331,7 @@ module FluxxRequestTransaction
               rtfs.updated_by_id = self.updated_by_id
               rtfs.save
             end
-          elsif rtfs
+          elsif rtfs && rtfs.id
             # The user removed the value, let's delete the record as well
             rtfs.destroy
           end
